Minor update to allow timezone offsets with decimals (e.g. for India - 5.5)

+++ src/IPGeolocation.cpp
@@ -67,7 +67,7 @@ void IPGeolocation::updateStatus(IPGeo *I){
   I->tz = doc["timezone"].as<String>();
   I->is_dst = doc["is_dst"];
   int dst_savings = doc["dst_savings"].as<int>();
-  int timezone_offset = doc["timezone_offset"];
+  double timezone_offset = doc["timezone_offset"];
   I->offset= ((I->is_dst) ? dst_savings : 0) + timezone_offset;
   I->country = doc["geo"]["country_name"].as<String>();
   I->country_code = doc["geo"]["country_code2"].as<String>();
